requirements.txt
pillow>=8.3.1
opencv-python>=3.4.1.15
scipy>=1.1.0
numpy>=1.19.5
torch>=1.5.0